[Refactor] Some refactoring to kernel.cpp
- Split KernelSource to a separate header
- Make use of file_paths when reading binaries

These changes turned out to be helpful when implementing the compile server.
Extracting them into a refactor PR.
